+++ client/public/index.html
@@ -24,6 +24,7 @@
   <!-- CSS  -->
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <link href="materializeCSS/materialize.css" type="text/css" rel="stylesheet" media="screen,projection" />
+  <link href="materializeCSS/materialize_style.css" type="text/css" rel="stylesheet" media="screen,projection" />
   <!--  Scripts-->
   <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
   <script src="materializeJS/materialize.js"></script>
