+++ client/public/materializeCSS/materialize.css
@@ -5795,7 +5795,7 @@ body.keyboard-focused .dropdown-content li:focus {
 }
 
 .modal .modal-footer .btn, .modal .modal-footer .btn-large, .modal .modal-footer .btn-small, .modal .modal-footer .btn-flat {
-  margin: 6px 0;
+  margin: 6px auto;
 }
 
 .modal-overlay {
