+++ client/public/materializeCSS/materialize_style.css
@@ -6,6 +6,7 @@
  * Made By MaterializeCSS.com
  */
 
+
  nav ul a,
  nav .brand-logo {
    color: #444;
