+++ client/src/components/Modals/index.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import Modal from 'react-bootstrap/Modal'
 import Button from 'react-bootstrap/Button'
-import ButtonToolbar from 'react-bootstrap/ButtonToolbar'
+// import "./style.css"
+
 
 class MyVerticallyCenteredModal extends React.Component {
     render() {
@@ -12,11 +13,6 @@ class MyVerticallyCenteredModal extends React.Component {
           aria-labelledby="contained-modal-title-vcenter"
           centered
         >
-          <Modal.Header closeButton>
-            <Modal.Title id="contained-modal-title-vcenter">
-              Modal heading
-            </Modal.Title>
-          </Modal.Header>
           <Modal.Body>
             <h4>Centered Modal</h4>
             <p>
@@ -25,40 +21,11 @@ class MyVerticallyCenteredModal extends React.Component {
               ac consectetur ac, vestibulum at eros.
             </p>
           </Modal.Body>
-          <Modal.Footer>
-            <Button onClick={this.props.onHide}>Close</Button>
+          <Modal.Footer closeButton>
+            <Button className="waves-effect waves-light" onClick={this.props.onHide}>Close</Button>
           </Modal.Footer>
         </Modal>
       );
     }
   }
-  
-  class App extends React.Component {
-    constructor(...args) {
-      super(...args);
-  
-      this.state = { modalShow: false };
-    }
-  
-    render() {
-      let modalClose = () => this.setState({ modalShow: false });
-  
-      return (
-        <ButtonToolbar>
-          <Button
-            variant="primary"
-            onClick={() => this.setState({ modalShow: true })}
-          >
-            Launch vertically centered modal
-          </Button>
-  
-          <MyVerticallyCenteredModal
-            show={this.state.modalShow}
-            onHide={modalClose}
-          />
-        </ButtonToolbar>
-      );
-    }
-  }
-
-  export default {MyVerticallyCenteredModal, App}
+  export default MyVerticallyCenteredModal;
