+++ client/src/components/Modals/style.css
@@ -1,0 +1,3 @@
+.modalButton {
+    float: right
+}
