+++ client/src/components/Nav/style.css
@@ -5,11 +5,3 @@
 #logo {
     height: 100%;
 }
-.btn {
-    height: 50%;
-    margin-top: 16px;
-    margin-right: 20px;
-    padding-right: 10px;
-    float: left;
-    text-align: center;
-}
